+++ src/bigxml/handler.py
@@ -11,7 +11,11 @@ _ATTR_MARKER = "_xml_handlers_on"
 
 def handler_from_dict(handler_dict):
     def handle(node):
-        handler = handler_dict.get(node.name)
+        handler = None
+        if hasattr(node, "namespace"):
+            handler = handler_dict.get("{{{}}}{}".format(node.namespace, node.name))
+        if handler is None:
+            handler = handler_dict.get(node.name)
         value = None
         if handler is not None:
             if isinstance(handler, dict):